EurekaImage-flow-interrupt-disable

+++ src/org/folio/eureka/EurekaImage.groovy
@@ -100,11 +100,16 @@ class EurekaImage implements Serializable {
                   it['id'] = "${pom.getArtifactId()}-${pom.getVersion()}" as String
                 }
               }
-              steps.writeJSON(file: "eureka-platform.json", json: eureka_platform, pretty: 0)
-              steps.sh(script: "mv eureka-platform.json data.json && jq '.' data.json > eureka-platform.json") // beatify JSON
-              steps.sh(script: "rm -f data.json && git commit -am '[EPL] updated: ${pom.getArtifactId()}-${pom.getVersion()}'")
-              steps.sh(script: "set +x && git push --set-upstream https://${steps.env.GIT_USER}:${steps.env.GIT_PASS}@github.com/folio-org/platform-complete.git snapshot")
-              logger.info("Snapshot branch successfully updated\n${moduleName} version: ${pom.getArtifactId()}-${pom.getVersion()}")
+              if (!("${pom.getArtifactId()}-${pom.getVersion()}" in eureka_platform['id'])) {
+                steps.writeJSON(file: "eureka-platform.json", json: eureka_platform, pretty: 0)
+                steps.sh(script: "mv eureka-platform.json data.json && jq '.' data.json > eureka-platform.json")
+                // beatify JSON
+                steps.sh(script: "rm -f data.json && git commit -am '[EPL] updated: ${pom.getArtifactId()}-${pom.getVersion()}'")
+                steps.sh(script: "set +x && git push --set-upstream https://${steps.env.GIT_USER}:${steps.env.GIT_PASS}@github.com/folio-org/platform-complete.git snapshot")
+                logger.info("Snapshot branch successfully updated\n${moduleName} version: ${pom.getArtifactId()}-${pom.getVersion()}")
+              } else {
+                logger.info("Version ${pom.getArtifactId()}-${pom.getVersion()} already exists.")
+              }
             }
           }
         }
